requirements.txt
streamlit
azure-ai-inference
azure-core
beautifulsoup4
requests